ignore files files which are not found by files() (#2538)
* rename lookup-files.spec.js to file-utils.spec.js

* .files should not choke on broken symlinks
